[win10] network: move implementation to platform folder
