+++ cmake/treedata/windowsstore/subdirs.txt
@@ -1,10 +1,10 @@
 xbmc/platform/win10             platform/win10
 xbmc/platform/win10/filesystem  platform/win10/filesystem
 xbmc/platform/win10/storage     platfrom/win10/storage
+xbmc/platform/win10/network     platform/win10/network
 xbmc/platform/win32/filesystem  platform/win32/filesystem
 xbmc/input/touch                input/touch
 xbmc/input/touch/generic        input/touch/generic
-xbmc/network/win10              network/win10
 xbmc/network/mdns               network/mdns
 xbmc/peripherals/bus/win10      peripherals/bus/win10
 xbmc/powermanagement/win10      powermanagement/win10
