+++ xbmc/platform/win10/network/CMakeLists.txt
@@ -2,4 +2,4 @@ set(SOURCES NetworkWin10.cpp)
 
 set(HEADERS NetworkWin10.h)
 
-core_add_library(network_win10)
+core_add_library(platform_win10_network)
