+++ xbmc/platform/win10/network/NetworkWin10.h
